{"schema":"decodex/commit/1","summary":"Extract rendering frame pipeline owner","authority":"manual"}
